Backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
celery
pika
transformers[sentencepiece]
requests
python-dotenv
